Implementação da stamina e scripts separados
correção de bugs, separação das features de movimento em cada script e criação de uma barra de estamina.
