+++ Projeto_2/Assets/Code/Scripts/Player/Player_Movement.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 79b86819fb7448044a921ab0bd0f875a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
