+++ Projeto_2/Assets/Code/Scripts/Player/Player_Movement/Dash.cs
@@ -1,0 +1,77 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Dash : MonoBehaviour
+{
+    public float dashSpeed; //dash speed of the player
+    public float dashTime; //dashing time
+    public float dashCooldown = 1f;  //time remaining to use it again
+    private float dashCounter; //state of time for dashing time
+    private float dashCoolCounter; //state of time for time remaining to use dash again
+    [SerializeField] private GameObject dashShieldImage;
+    [SerializeField] TrailRenderer tr; //trail render to design impulse
+    public bool dashShieldPlayer;
+
+    public PlayerMovement movement;
+    public Shield shield;
+
+    // Start is called before the first frame update
+    void Start()
+    {
+        dashShieldPlayer = false;
+        tr.emitting = false; 
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        if(Input.GetKeyDown(KeyCode.Space)) //if space is pressed...
+        {//and the cooldown and dash time is equal to zero: player speed changes to dash, trail render is activated and the state
+        // of time of dash start the countdown to reset
+            if(dashCoolCounter <= 0 && dashCounter <= 0 && shield.shieldPlayer == false)
+            {
+                if(movement.moveInput.y == 0f)
+                {
+                    movement.activeMoveSpeed = dashSpeed;
+                    movement.rb.constraints = RigidbodyConstraints2D.FreezePositionY | RigidbodyConstraints2D.FreezeRotation ;
+                    dashShieldImage.SetActive(true);
+                    dashShieldPlayer = true;
+                    tr.emitting = true;
+                    dashCounter = dashTime;   
+                }
+
+                if(movement.moveInput.x == 0f)
+                {
+                    movement.activeMoveSpeed = dashSpeed;
+                    movement.rb.constraints = RigidbodyConstraints2D.FreezePositionX | RigidbodyConstraints2D.FreezeRotation;
+                    dashShieldImage.SetActive(true);
+                    dashShieldPlayer = true;
+                    tr.emitting = true;
+                    dashCounter = dashTime;
+                }
+        
+            }
+        }
+
+        if(dashCounter > 0) //countdown of dashing time is reseting
+        {
+            dashCounter -= Time.deltaTime;
+ //but, the countdown is already reset(equal to zero): player speed changes to normal speed, trail render is deactivated and cooldown start its countdown
+            if(dashCounter <= 0)
+            {
+                movement.activeMoveSpeed = movement.moveSpeed;
+                movement.rb.constraints = RigidbodyConstraints2D.FreezeRotation;
+                dashShieldImage.SetActive(false);
+                dashShieldPlayer = false;
+                tr.emitting = false;
+                dashCoolCounter = dashCooldown;
+            }
+        }
+
+        if(dashCoolCounter > 0) //countdown of cooldown is reseting
+        {
+            dashCoolCounter -= Time.deltaTime;
+        }
+    }
+}
