+++ Projeto_2/Assets/Code/Scripts/Player/Player_Movement/Dash.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 92c4ff19abf7ffa43ad193a612ab8adc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
