+++ Projeto_2/Assets/Code/Scripts/Player/Player_Movement/Shield.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f77209ba4d38fc64f8925a0c660ee995
+guid: 20104b2416bcaa34196a58d216400d82
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
