+++ Projeto_2/Assets/Code/Scripts/Player/Stamina.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1b5bea8bce0878f49a513e33fbbdfe2f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
